Remove StorageHelper in favour of FileRepository
